pyproject.toml
[tool.poetry]
name = "fakegoodsmile"
version = "0.1.0"
description = ""
authors = ["Joaquin Huaman Pino <quinino1@gmail.com>"]
readme = "README.md"

[tool.poetry.dependencies]
python = "^3.11"
django = "^4.2.3"
djangorestframework = "^3.14.0"
python-dotenv = "^1.0.0"
pytest-django = "^4.5.2"
psycopg2-binary = "^2.9.6"
pillow = "^10.0.0"
coreapi = "^2.3.3"
baker = "^1.3"
model-bakery = "^1.14.0"
djangorestframework-simplejwt = "^5.3.0"


[tool.poetry.group.dev.dependencies]
pylint = "^2.17.4"
pylint-django = "^2.5.3"
black = "^23.7.0"

[build-system]
requires = ["poetry-core"]
build-backend = "poetry.core.masonry.api"


[tool.pytest.ini_options]
log_cli = true
log_cli_level = "CRITICAL"
log_cli_format = "%(message)s"

log_file = "logs/pytest.log"
log_file_level = "DEBUG"
log_file_format = "%(asctime)s [%(levelname)8s] %(message)s (%(filename)s:%(lineno)s)"
log_file_date_format = "%Y-%m-%d %H:%M:%S"